+++ src/gui/enums.rs
@@ -1,26 +1,8 @@
-use strum_macros::EnumString;
-
 #[derive(Clone, Copy)]
 pub enum BaseColor {
 	Red,
 	Green,
 	Blue,
-}
-
-#[derive(Clone, Copy, EnumString)]
-pub enum Effects {
-	Static,
-	Breath,
-	Smooth,
-	LeftWave,
-	RightWave,
-	Lightning,
-	AmbientLight,
-	SmoothLeftWave,
-	SmoothRightWave,
-	LeftSwipe,
-	RightSwipe,
-	Disco,
 }
 
 pub enum Colors {
@@ -33,15 +15,4 @@ pub enum Colors {
 	LightGray = 0x777777,
 	LighterGray = 0xcccccc,
 	White = 0xffffff,
-}
-
-pub enum Message {
-	UpdateAllValues { value: [f32; 12] },
-	UpdateRGB { index: u8, value: f32 },
-	UpdateZone { zone_index: u8, value: [f32; 3] },
-	UpdateEffect { effect: Effects },
-	UpdateValue { index: u8, value: f32 },
-	UpdateBrightness { brightness: u8 },
-	UpdateSpeed { speed: u8 },
-	Restart,
 }
