+++ src/gui/mod.rs
@@ -2,5 +2,4 @@ pub mod builder;
 mod effect_browser_tile;
 pub mod enums;
 mod keyboard_color_tiles;
-pub mod keyboard_manager;
 mod options_tile;
